optimized upload button front end
